backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
sqlmodel
psycopg2
python-dotenv
httpx
boto3